+++ app/src/main/java/com/application/traverldiary/models/Journey.kt
@@ -2,7 +2,7 @@ package com.application.traverldiary.models
 
 open class Journey(
     var nameOrId: String,
-    var fromTime: Int,
-    var toTime: Int,
+    var fromTime: Float,
+    var toTime: Float,
     var isHappened: Boolean = false
 )
